+++ src/components/App/App.tsx
@@ -9,7 +9,6 @@ import {OnboardingPageProvider} from '../Providers/OnboardingPageProvider';
 import OnboardingScreen from "./OnboardingScreen/OnboardingScreen";
 import UnlockScreen from "./UnlockScreen/UnlockScreen";
 import LaunchScreen from "./LaunchScreen/LaunchScreen";
-import {deletePasscode, initializeWorkerService, isVerified} from "../../services/workerService/workerService";
 
 const StyledContainer = styled.div`
   height: 100vh;
@@ -43,6 +42,7 @@ const App = () => {
     }
 
     useEffect(() => {
+
 
         initialize().then(() => {
         })
@@ -91,16 +91,11 @@ const App = () => {
                 return (
                     <MainScreen/>
                 )
-            case hasContainer && !isUnlocked && isVerified():
+            case hasContainer && !isUnlocked:
                 return (
                     <UnlockScreen/>
                 )
-            case (!hasContainer && !isUnlocked) || !isVerified():
-                deletePasscode().then((status) => {
-                    if (status === 'SUCCESS') {
-                        initializeWorkerService()
-                    }
-                })
+            case !hasContainer && !isUnlocked:
                 return (
                     <OnboardingPageProvider
                         existingPages={['language', 'setPasscode', 'confirmPasscode', 'verification', 'settingUp']}>