+++ src/components/App/OnboardingScreen/SettingUpPage/SettingUpPage.tsx
@@ -17,9 +17,9 @@ import AlertDialog, {AlertDialogActions} from "../../../UI/Common/AlertDialog/Al
 import {Warning} from "../../../UI/Icons/Icons";
 import {
     createPasscode,
-    getWorkerService, hasPasscode,
-    Preferences,
-    savePreferences, verifyInvitation
+    hasPasscode,
+    savePreferences,
+    verifyInvitation
 } from "../../../../services/workerService/workerService";
 
 const StyledContainer = styled.div`
@@ -106,8 +106,8 @@ const SettingUpPage = () => {
     const {state, dispatch} = useOnboardingPageNavigator()
 
     const setupStateText = [
-        'Creating container',
-        'Verifying invitation code'
+        <FormattedMessage id='onboarding.setup.create.container'/>,
+        <FormattedMessage id='onboarding.setup.verify.code'/>
     ]
 
     const [setupState, setSetupState] = useState<number>(1)
@@ -168,7 +168,9 @@ const SettingUpPage = () => {
                         }
                     }).then((status) => {
                         if (status === "SUCCESS") {
-                            setSetupState(prevState => prevState + 1)
+                            savePreferences({theme: theme, language: locale}).then((status) => {
+                                setSetupState(prevState => prevState + 1)
+                            })
                         }
                     })
                 } else {
@@ -183,30 +185,10 @@ const SettingUpPage = () => {
                 }
                 // Verification
                 break;
-            case 3:
-                // createPasscode({
-                //     passcode: state.onboardingPageData.passcode, progress: () => {
-                //     }
-                // }).then((status) => {
-                //     if (status === "SUCCESS") {
-                //         storePreferences().then(() => {
-                //             dispatch(onboardingActions.setVerificationStatus('SUCCESS'))
-                //         })
-                //     }
-                // })
-                break;
             default:
                 break;
         }
     }, [setupState])
-
-    const storePreferences = () => {
-        const preferences: Preferences = {
-            theme: theme,
-            language: locale
-        }
-        return savePreferences(preferences)
-    }
 
     const onSetupComplete = () => {
         setHasContainer(true)
@@ -289,7 +271,7 @@ const SettingUpPage = () => {
             </StyledPageUpper>
             <StyledPageLower>
                 {state.onboardingPageData?.verificationStatus === 'SUCCESS'
-                    ? <Button onClick={onSetupComplete}>Enter Seguro</Button>
+                    ? <Button onClick={onSetupComplete}><FormattedMessage id='onboarding.setup.enter.button'/></Button>
                     : <ProgressSteps
                         currentStage={setupState}
                         numberOfSteps={setupStateText.length}