+++ src/components/UI/LanguageSelect/Language/Language.tsx
@@ -20,7 +20,7 @@ const StyledItem = styled.li<{ selected: boolean }>`
   cursor: pointer;
   background-color: ${props => props.selected ?
           props.theme.ui.colors.primary :
-          props.theme.ui.colors.background.foundation};
+          "transparent"};
   color: ${props => props.selected ? 'white' : props.theme.ui.colors.text.primary};
   font-size: ${props => props.theme.ui.fontSizes.narrow.sm};
   //&:first-of-type {