+++ src/electron/electron.js
@@ -1,5 +1,5 @@
-const { createClientServer } = require('./server')
-const { createWindow } = require('./window')
+const {createClientServer} = require('./server')
+const {createWindow} = require('./window')
 
 const isDevelopmentMode = process.env.NODE_ENV === 'development'
 
@@ -10,14 +10,13 @@ const isDevelopmentMode = process.env.NODE_ENV === 'development'
 
     if (isDevelopmentMode) {
         console.log(`connecting to client dev server on port ${clientServerPort}`)
-    }
-    else {
+    } else {
         console.log('creating client server')
         clientServerPort = (await createClientServer()).clientServerPort
     }
 
     console.log('creating electron window')
-    await createWindow({ clientServerPort })
+    await createWindow({clientServerPort})
     console.log('created electron window')
 
     console.log('application booted')