+++ src/localization/messages/en-US.ts
@@ -31,6 +31,10 @@ const en_US: Messages = {
     'onboarding.verification.modal.button.update': 'Update',
     'onboarding.verification.modal.button.newCode': 'Enter new verification code',
 
+    'onboarding.setup.create.container': 'Creating container...',
+    'onboarding.setup.verify.code': 'Verifying invitation code...',
+    'onboarding.setup.enter.button': 'Enter Seguro',
+
     'onboarding.carousel.title.seguro-platform': 'Seguro Platform',
     'onboarding.carousel.title.no-ip-tracking': 'No IP tracking',
     'onboarding.carousel.title.decentralized': 'Decentralized',
@@ -50,6 +54,11 @@ const en_US: Messages = {
 
     'unlock.title': 'Enter your Seguro passcode',
 
+    'platform.unlock.button.forgot': 'Forgot your passcode?',
+    'platform.dialog.delete.message': 'Deleting your container is a permanent action, you will need to create a new passcode with a new invitation code.',
+    'platform.dialog.delete.button.cancel': 'Cancel',
+    'platform.dialog.delete.button.confirm': 'Delete',
+
     'keypad.cancel': 'Cancel',
     'keypad.unlock': 'Unlock',
 
@@ -63,6 +72,10 @@ const en_US: Messages = {
 
     'platform.overlay.unlocking': 'Unlocking platform...',
     'platform.overlay.createProfile': 'Creating profile...',
+
+    'platform.modal.verification.title': 'Verification Required',
+    'platform.modal.verification.text': 'Please enter your 36 character invitation code to complete your setup.',
+    'platform.modal.verification.button': "Verify",
 
     'platform.network.status.label': 'Current Status:',
     'platform.network.status.low': 'Low',